move repl code to replication.h
